Move PotelSentrySpanProcessor to SentrySpanProcessor
* remove the older one
* fix propagator `inject` to not rely on old SentrySpanProcessor span
  map
