+++ sentry_sdk/integrations/opentelemetry/propagator.py
@@ -1,7 +1,10 @@
+from typing import cast
+
 from opentelemetry import trace
 from opentelemetry.context import (
     Context,
     get_current,
+    get_value,
     set_value,
 )
 from opentelemetry.propagators.textmap import (
@@ -21,9 +24,7 @@ from opentelemetry.trace import (
 from sentry_sdk.integrations.opentelemetry.consts import (
     SENTRY_BAGGAGE_KEY,
     SENTRY_TRACE_KEY,
-)
-from sentry_sdk.integrations.opentelemetry.span_processor import (
-    SentrySpanProcessor,
+    SENTRY_SCOPES_KEY,
 )
 from sentry_sdk.tracing import (
     BAGGAGE_HEADER_NAME,
@@ -35,6 +36,7 @@ from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
     from typing import Optional, Set
+    from sentry_sdk.integrations.opentelemetry.scope import PotelScope
 
 
 class SentryPropagator(TextMapPropagator):
@@ -47,6 +49,7 @@ class SentryPropagator(TextMapPropagator):
         if context is None:
             context = get_current()
 
+        # TODO-neel-potel cleanup with continue_trace / isolation_scope
         sentry_trace = getter.get(carrier, SENTRY_TRACE_HEADER_NAME)
         if not sentry_trace:
             return context
@@ -89,27 +92,15 @@ class SentryPropagator(TextMapPropagator):
         if context is None:
             context = get_current()
 
-        current_span = trace.get_current_span(context)
-        current_span_context = current_span.get_span_context()
+        scopes = get_value(SENTRY_SCOPES_KEY, context)
+        if scopes:
+            scopes = cast("tuple[PotelScope, PotelScope]", scopes)
+            (current_scope, _) = scopes
 
-        if not current_span_context.is_valid:
-            return
-
-        span_id = trace.format_span_id(current_span_context.span_id)
-
-        span_map = SentrySpanProcessor().otel_span_map
-        sentry_span = span_map.get(span_id, None)
-        if not sentry_span:
-            return
-
-        setter.set(carrier, SENTRY_TRACE_HEADER_NAME, sentry_span.to_traceparent())
-
-        if sentry_span.containing_transaction:
-            baggage = sentry_span.containing_transaction.get_baggage()
-            if baggage:
-                baggage_data = baggage.serialize()
-                if baggage_data:
-                    setter.set(carrier, BAGGAGE_HEADER_NAME, baggage_data)
+            # TODO-neel-potel check trace_propagation_targets
+            # TODO-neel-potel test propagator works with twp
+            for (key, value) in current_scope.iter_trace_propagation_headers():
+                setter.set(carrier, key, value)
 
     @property
     def fields(self):
